Move discon controller files under src/utilities/controllers
