+++ src/utilities/controllers/discon.h
@@ -1,22 +1,3 @@
-//**********************************************************************************************************************************
-// LICENSING
-// Copyright (C) 2015-2016  National Renewable Energy Laboratory
-// Copyright (C) 2016-2017  Envision Energy USA, LTD
-// Copyright (C) 2023       National Renewable Energy Laboratory
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//**********************************************************************************************************************************
-
 #pragma once
 
 #ifdef __cplusplus
