+++ tests/unit_tests/CMakeLists.txt
@@ -9,7 +9,7 @@ target_sources(
     PRIVATE
     utest_main.cpp
     test_config.cpp
-    ../../src/utilities/discon.c
+    ../../src/utilities/controllers/discon.c
 )
 
 target_compile_options(
